wait till all candidates are folded

+++ pipelines/trapum_fold_and_score_pipeline/webpage_fold_and_score.py
@@ -183,38 +183,39 @@ def extract_fold_and_score(processing_args,processing_id,output_dir,xml_file,dp_
     extra = no_of_cands%batch_no
     batches = int(no_of_cands/batch_no) +1
     for x in range(batches):
-       start = x*batch_no
-       if(x==batches-1):
-           end = x*batch_no+extra
-       else:
-           end = (x+1)*batch_no   
-       for i in range(start,end):
-           folding_packet={}
-           folding_packet['period'] = mod_period[i]
-           folding_packet['acc'] = acc[i]
-           folding_packet['pdot'] = pdot[i] 
-           folding_packet['dm'] = dm[i] 
-           output_name= "dm_%.2f_acc_%.2f_candidate_number_%d"%(folding_packet['dm'],folding_packet['acc'],i)
-           try:
-               process = subprocess.Popen("prepfold -ncpus 1 -mask %s -noxwin -topo -p %s -pd %s -dm %s %s -o %s"%(mask_path,str(folding_packet['period']),str(folding_packet['pdot']),str(folding_packet['dm']),input_name,output_name),shell=True,cwd=output_path)
-           except Exception as error:
-               log.error(error)
+        start = x*batch_no
+        if(x==batches-1):
+            end = x*batch_no+extra
+        else:
+            end = (x+1)*batch_no   
+        for i in range(start,end):
+            folding_packet={}
+            folding_packet['period'] = mod_period[i]
+            folding_packet['acc'] = acc[i]
+            folding_packet['pdot'] = pdot[i] 
+            folding_packet['dm'] = dm[i] 
+            output_name= "dm_%.2f_acc_%.2f_candidate_number_%d"%(folding_packet['dm'],folding_packet['acc'],i)
+            try:
+                process = subprocess.Popen("prepfold -ncpus 1 -mask %s -noxwin -topo -p %s -pd %s -dm %s %s -o %s"%(mask_path,str(folding_packet['period']),str(folding_packet['pdot']),str(folding_packet['dm']),input_name,output_name),shell=True,cwd=output_path)
+            except Exception as error:
+                log.error(error)
  
-       if  process.communicate()[0]==None:
-           continue
-       else:
-           time.sleep(10)
+        if  process.communicate()[0]==None:
+            continue
+        else:
+            time.sleep(60)
 
-    log.info("Folding done for processing. Scoring all candidates...") 
-
-    tar_name = subprocess.getoutput("python2 webpage_score.py --in_path=%s"%output_path)
-    
-    log.info("Scoring done...")     
- 
-    # Remove original files 
-    subprocess.check_call("rm *pfd* *.txt",shell=True,cwd=output_path)
-
-    return os.path.basename(tar_name),output_path
+    while True:
+        if len(glob.glob('%s/*.pfd'%output_path)) == no_of_cands:
+            log.info("Folding done for all candidates. Scoring all candidates...") 
+            tar_name = subprocess.getoutput("python2 webpage_score.py --in_path=%s"%output_path)
+            log.info("Scoring done...")     
+            # Remove original files 
+            subprocess.check_call("rm *pfd* *.txt",shell=True,cwd=output_path)
+            return os.path.basename(tar_name),output_path
+        else:
+            log.info("Still not finished folding...")
+            time.sleep(60)     
 
                    
 if __name__ == '__main__':
